+++ program_structure/src/abstract_syntax_tree/ast.rs
@@ -1,4 +1,4 @@
-use crate::{file_definition::{FileLocation, FileID}, error_definition::Report, error_code::{ReportCode, produce_report}};
+use crate::{file_definition::{FileLocation, FileID}, error_definition::Report, error_code::{ReportCode}};
 use num_bigint::BigInt;
 use serde_derive::{Deserialize, Serialize};
 
@@ -460,4 +460,151 @@ impl MemoryKnowledge {
             panic!("abstract memory address was look at without being initialized");
         }
     }
+}
+
+ pub fn produce_report(error_code: ReportCode, location : FileLocation, file_id : FileID) -> Report {
+    use ReportCode::*;
+    let report  = match error_code {
+            UnclosedComment => {
+                let mut report =
+                    Report::error("unterminated /* */".to_string(), ReportCode::UnclosedComment);
+                report.add_primary(location, file_id, "Comment starts here".to_string());
+                report
+            }
+            NoMain => Report::error(
+                "No main specified in the project structure".to_string(),
+                ReportCode::NoMainFoundInProject,
+            ),
+            MultipleMain =>{
+                Report::error(
+                    "Multiple main components in the project structure".to_string(),
+                    ReportCode::MultipleMainInComponent,
+                )
+            }
+            MissingSemicolon => {
+                let mut report = Report::error(format!("Missing semicolon"), 
+                    ReportCode::MissingSemicolon);
+                report.add_primary(location, file_id, "A semicolon is needed here".to_string());
+                report
+            }
+            UnrecognizedInclude => {
+                let mut report =
+                Report::error("unrecognized argument in include directive".to_string(), ReportCode::UnrecognizedInclude);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+
+            }
+            UnrecognizedPragma => {
+                let mut report =
+                Report::error("unrecognized argument in pragma directive".to_string(), ReportCode::UnrecognizedPragma);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+
+            }        
+            UnrecognizedVersion => {
+                let mut report =
+                Report::error("unrecognized version argument in pragma directive".to_string(), ReportCode::UnrecognizedVersion);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+            }      
+            IllegalExpression => {
+                let mut report =
+                Report::error("illegal expression".to_string(), ReportCode::IllegalExpression);
+            report.add_primary(location, file_id, "here".to_string());
+            report
+            }
+            MultiplePragma => {
+                let mut report =
+                Report::error("Multiple pragma directives".to_string(), ReportCode::MultiplePragma);
+            report.add_primary(location, file_id, "here".to_string());
+            report
+            },
+            _ => unreachable!(),    
+    };
+    report
+}
+
+pub fn produce_version_warning_report(path : String, version : Version) -> Report {
+    Report::warning(
+        format!(
+            "File {} does not include pragma version. Assuming pragma version {:?}",
+            path, version
+        ),
+        ReportCode::NoCompilerVersionWarning,
+    )
+}
+
+
+pub fn produce_report_with_message(error_code : ReportCode, msg : String) -> Report {
+    match error_code {
+        ReportCode::FileOs => {
+            Report::error(
+            format!("Could not open file {}", msg),
+            ReportCode::FileOs,
+            )
+        }
+        ReportCode::IncludeNotFound => {
+            Report::error(
+                format!(" The file {} to be included has not been found", msg),
+                ReportCode::FileOs,
+                )
+        },
+        _ => unreachable!()
+    }
+}
+
+pub fn produce_generic_report( msg : String, location : FileLocation, file_id : FileID) -> Report {
+    let mut report = Report::error(msg, ReportCode::GenericParsing);
+    report.add_primary(location, file_id, "Error here".to_string());
+    report
+}
+pub fn produce_compiler_version_report(path : String, required_version : Version, version :  Version) -> Report {
+    let report = Report::error(
+        format!("File {} requires pragma version {:?} that is not supported by the compiler (version {:?})", path, required_version, version ),
+        ReportCode::CompilerVersionError,
+    );
+    report
+}
+
+pub fn anonymous_inside_condition_error(meta : Meta) -> Report {
+    let msg = "An anonymous component cannot be used inside a condition ".to_string();
+                let mut report = Report::error(
+                    format!("{}", msg),
+                    ReportCode::AnonymousCompError,
+                );
+                let file_id = meta.get_file_id().clone();
+                report.add_primary(
+                    meta.location,
+                    file_id,
+                    "This is an anonymous component used inside a condition".to_string(),
+                );
+                report
+}
+
+pub fn anonymous_general_error(meta : Meta, msg : String) -> Report {
+    let mut report = Report::error(
+                    format!("{}", msg),
+                    ReportCode::AnonymousCompError,
+                );
+                let file_id = meta.get_file_id().clone();
+                report.add_primary(
+                    meta.location,
+                    file_id,
+                    "This is the anonymous component whose use is not allowed".to_string(),
+                );
+                report
+}
+
+pub fn tuple_general_error(meta : Meta, msg : String) -> Report {
+    let mut report = Report::error(
+                    format!("{}", msg),
+                    ReportCode::TupleError,
+                );
+                let file_id = meta.get_file_id().clone();
+                report.add_primary(
+                    meta.location,
+                    file_id,
+                    "This is the tuple whose use is not allowed".to_string(),
+                );
+                report
 }
